Migrate accounts_config to TypeScript and refine accounts connection logic
• Replace accounts_config.js with a new TypeScript implementation in accounts_config.ts
• Update import references in accounts.ts, core.js, and AccountsAddEdit.svelte to use the .ts file
• Refine the setupPing logic in accounts.ts to properly handle missing timestamps and reconnection conditions
