+++ src/core/core.js
@@ -11,7 +11,7 @@ import { accounts_init, accounts, active_account_store, active_account, selectAc
 export { accounts, active_account_store, active_account, selectAccount, active_account_module_data, findAccount };
 
 // Import accounts config
-import { accounts_config, findAccountConfig, accountConfigExistsByCredentials } from './accounts_config.js';
+import { accounts_config, findAccountConfig, accountConfigExistsByCredentials } from './accounts_config.ts';
 export { accounts_config, findAccountConfig, accountConfigExistsByCredentials };
 export const accountConfigExistsById = id => !!findAccountConfig(id);
